+++ src/Navigation/SideNavi.jsx
@@ -1,39 +1,27 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
-import { Menu } from 'antd';
 import { withRouter } from 'react-router';
-import { getActiveRoutes } from './routing';
 import { renderMenu } from './menu';
-
-const StyledMenu = styled(Menu)`
-    background: transparent !important;
-    line-height: 64px !important;
-    display: inline-block;
-`;
+import {Navigation} from "./Navigation";
 
 export const SideNavi = withRouter((props) => {
-    const {routes, location, defaultOpenKeys} = props;
-
-    const activeRoutes = getActiveRoutes(routes, location);
+    const {routes} = props;
 
     return (
-        <StyledMenu
+        <Navigation
             mode="inline"
-            selectedKeys={activeRoutes}
-            defaultOpenKeys={defaultOpenKeys}
+            routes={routes}
         >
             {renderMenu(routes)}
-        </StyledMenu>
+        </Navigation>
     );
 });
 
 SideNavi.defaultProps = {
     routes: [],
-    defaultOpenKeys: []
 };
 
 SideNavi.propTypes = {
     routes: PropTypes.arrayOf(PropTypes.object),
-    defaultOpenKeys: PropTypes.arrayOf(PropTypes.string)
 };
+
